pyproject.toml
[project]
name = "guardrailed-chatbot"
version = "0.1.0"
description = "Didactic example: conversational chatbot with input/output guardrails"
requires-python = ">=3.13"
dependencies = [
    # Official Google Gemini SDK
    "google-genai>=1.10.0",
    # Official Python client for Ollama (local model server)
    "ollama>=0.4.0",
    # Loads variables from the .env file into os.environ
    "pydantic-settings>=2.13.1",
    "python-dotenv>=1.1.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.3",
    "pytest-mock>=3.15.1",
]
